pyproject.toml
[tool.poetry]
name = "ejercicioia6"
version = "0.1.0"
description = ""
authors = ["Jose Antonio <jaborfer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.9"
langchain-community = "^0.3.7"
pydf = "^12"
langchain-chroma = "^0.1.4"
pypdf = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
